refactored utilitary to get language
